refactor: unbundle plugin file
