+++ inlang/source-code/sdk2/src/lix-plugin/utilities/getLastChildOfChange.ts
